requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.28.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
fastapi==0.103.2
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
packaging==23.2
pip==22.3.1
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
setuptools==65.5.0
sniffio ==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
alembic==1.12.0
greenlet==3.0.0
Mako==1.2.4
MarkupSafe==2.1.3
pip==22.3.1
psycopg2==2.9.9
setuptools==65.5.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
